🐛fix

+++ src/interactive/repl.ts
@@ -58,17 +58,19 @@ async function confirmKill() {
         return false
     }
 }
-async function stopREPL() {
+async function stopREPL(onDeactivate=false) {
     const config = vscode.workspace.getConfiguration('julia')
     if (Boolean(config.get('persistentSession.enabled'))) {
-        try {
-            const sessionName = parseSessionArgs(config.get('persistentSession.tmuxSessionName'))
-            const killSession = await confirmKill()
-            if (killSession) {
-                await exec(`tmux kill-session -t ${sessionName}`)
+        if (!onDeactivate) {
+            try {
+                const sessionName = parseSessionArgs(config.get('persistentSession.tmuxSessionName'))
+                const killSession = await confirmKill()
+                if (killSession) {
+                    await exec(`tmux kill-session -t ${sessionName}`)
+                }
+            } catch (err) {
+                vscode.window.showErrorMessage('Failed to close tmux session: ' + err.stderr)
             }
-        } catch (err) {
-            vscode.window.showErrorMessage('Failed to close tmux session: '+err.stderr)
         }
     }
     if (isConnected()) {
@@ -1263,5 +1265,5 @@ export function activate(context: vscode.ExtensionContext, compiledProvider, jul
 }
 
 export function deactivate() {
-    stopREPL()
+    stopREPL(true)
 }
